Type cast size_t to int as needed. size_t is U64 on windows 64 bit machines. GCC defines size_t as int. It is reasonable to type definitions as size_t when that are relating to the sizeof() elements. For portability with windows casting size_t to int should be alright since most of the left-sided elements are usually int.
git-svn-id: file:///home/jj/hercules.svn/trunk@6034 956126f8-22a0-4046-8f4a-272fa8102e63

+++ tapeccws.c
@@ -3222,7 +3222,7 @@ static BYTE     write_immed    = 0;     /* Write-Immed. mode active  */
             int  i;   // (work)
 
             /* Calculate residual byte count */
-            RESIDUAL_CALC (strlen(dev->filename));
+            RESIDUAL_CALC ((int)strlen(dev->filename));
 
             /* Copy device filename to guest storage */
             for (i=0; i < num; i++)
